tested the player motion code
It works. had to fill in a method I forgot was only stubbed and learned
something interesting.

An RPC call uses the first method with a given name it sees. That means,
basically, if you have one version of an function meant for RPCs and one
not, put the non-RPC one second, otherwise unitys says, "I see the
function, but its not an RPC. I give up".